Add a unit of work strategy.
The current twitter implementation resolves dependencies in parallel.
With this strategy, the same dependency can be resolved "at the same time" by different package instances.
When this happens, there was no control which result in I/O errors and slower dep tracking.
Now all the packages share a common unit of work that is a simple storage with lock ability.

Also fix CS.