Add Github Enterprise functionality to the GithubResolver
Adds the ability for a user to add the "github-enterprise" key to
their .bowerrc file.  If this key is present, it will also apply the
GithubResolver to any bower packages hosted at a repository on that
host.  Archives will be downloaded from these sources instead of a git
checkout.

Additionally this works around a bug that exists in recent versions of
Github Enterprise installations.  These installations neither support
shallow clones over https, nor do they return an error when one is
attempted.  As a result these clones hang forever and the `bower install`
step never completes. This commit will skip adding the `--depth 1` flag
to a clone command if cloning over https to a registered Github
Enterprise instance.