Changed smart Git host detection
Only check the first Content-Type returned by the server. Some hosts are
returning text/plain in the first response, and then the correct content
type in later responses. This seems to be the case for GitHub
Enterprise. By checking the first content type entry, we should be able
to verify this.

Also added unit tests to test this behavior.