Update README.md
Hi,

I added to the section "Support" a link to the Front-End Tooling with Gulp, Bower, and Yeoman  written by Stefan Baumgartner for Manning Publications. We believe this book is a great resource and all members of community will benefit from this information. If you find this inappropriate, please let me know where I can publish it.