Update `bower version` tests, refs #2139.
`bower version` currently takes the version from the "version" field of
bower.json, but the spec says that bower ignores it.

This commit updates the `bower version` tests so that they are
consistent with the spec. Now, the tests are based on the assumption
that `bower version` should only use git tags.