add support for save and save-exact in .bowerrc
The install command now supports designating "save" and "save-exact"
values in .bowerrc. If the `--save` flag is not used, but the config
file sets "save" to true, the command will behave as if the `--save`
option was specified. The same is true of `--save-exact` if
"save-exact" is set to true in `.bowerrc`.

The uninstall command will also behave as if the `--save` flag had
been specified if "save" is set to true in `.bowerrc`.