Do not store resolutions with force-latest when --dont-save-resolutions is used
Closes #2344

Without this, there's no way to prevent --force-latest from modifying bower.json when it resolves a conflict.
This adds a --dont-save-resolutions to prevent that.

Not sure this is the best way, but it's the compatible way :).

Also, I have no idea how to get at the parsed options in Manager, directly looking at cooked argv is probably not the correct way..