Prefer exact versions when dissecting dependencies
When a project has multiple dependencies on the same package, we try to
find one that resolves to a version that also satisfies all of the other
constraints. Dependencies are checked in the order that they were
resolved, which is dependent on network latency and essentially random.
If there are multiple valid resolutions, running `bower install` with
the same `bower.json` file won't always produce the same result, which
can lead to inconsistent behaviour across environments.

If one of the dependencies specifies an exact version, we should prefer
it over others that specify a range of versions. This is consistent with
how we already deprioritise wildcard targets - we want to choose the
most specific target of the highest version which resolves correctly.